If we have a component and container with the same name, ie it's just wiring it up, then merge into a single container. I think this avoids weird confusions of components & containers that are basically the same element when the component is only ever used in a redux context
